Add files via upload
The HR Analytics case study is being carried for the company who wants to act upon the cause of attrition in the company. Attrition in any company, be it because of the employees leaving voluntarily or being fired harms the companies reputation and the issues must be addressed in a right way.